split visitor
